requirements.txt
python-telegram-bot>=21.0
google-generativeai>=0.5.0
playwright>=1.40.0
python-dotenv>=1.0.0
